+++ frontend/controllers/analytics/position_new_controller.js
@@ -58,8 +58,6 @@ export default class extends ApplicationController {
       'PositionsReflex#create',
       {
         portfolio_id:   document.getElementById("portfolio_id").value,
-        show_plans:     document.getElementById("show_plans").value,
-        show_dividents: document.getElementById("show_dividents").value,
         locale:         document.getElementById("current_locale").value
       }
     )