+++ frontend/controllers/analytics/positions_controller.js
@@ -8,8 +8,6 @@ export default class extends ApplicationController {
       'PositionsReflex#index',
       {
         portfolio_id:   document.getElementById('portfolio_id').value,
-        show_plans:     document.getElementById('show_plans').value,
-        show_dividents: document.getElementById('show_dividents').value,
         locale:         document.getElementById('current_locale').value
       }
     )