+++ frontend/controllers/analytics_controller.js
@@ -9,7 +9,6 @@ export default class extends ApplicationController {
       {
         position_id:  positionIndex,
         portfolio_id: document.getElementById('portfolio_id').value,
-        show_plans:   document.getElementById('show_plans').value,
         locale:       document.getElementById('current_locale').value
       }
     )